Allow spaces in argument values.
This is a hack to allow spaces in argument values (for example, for
arguments like --title). The hack is to have the module substitute all
occurrences of tilde in the URI with spaces before passing the
arguments to rrdtool.

It comes at the cost of losing the ability to place literal tilde
characters in graph text, which is not likely to be a problem in
practice. If it is found to be a problem, perhaps the character to
substitue (or the feature altogether) could be controlled with a
module directive.